fix(cluster): set `Upsert: false` when creating cluster
All other resources managed by this provider have `Upsert: false` when creating.  This is because we want an error to occur if a duplicate cluster is created otherwise we can end up with two different Terraform resources managed the same resource in ArgoCD. Existing clusters must be imported into Terraform state.

Change to set Upsert = true was introduced in https://github.com/oboukili/terraform-provider-argocd/pull/75. PR is in general unrelated to this functionality.