fix: allow templated label values (#634)
`validateMetadataLabels` ensures that labels are in a certain format.
This works fine in most cases, except when this is used within appsets,
which allow for labels to be templated. This adds a layer of
indirection allowing for templated labels to be used, but only when an
appset makes use of this feature.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>