chore: add possibility to use testcontainers
As a convenience, it's nice to be able to run acceptance tests directly
without needing to setup the test environment first. While it's not
_too_ burdensome, I'm kinda lazy and prefer having a single step to do
this, and it's equally nice to not to have to pollute my own environment
with yet another k8s cluster configuration.

This change is fully compatible with the existing offering and needs to
be opted into by setting `USE_TESTCONTAINERS=true` when running `go test`
(or by using the `make testacc_testcontainers` task).

As a part of this, I needed to rework the kustomizations slightly so
that it's slightly more static (we now have overlays based on Argo CD
versions), instead of relying on needing to pull the Argo CD manifests
using curl.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>