chore: preload argocd before running testcontainer tests
While running this locally I got a few tests that were persistently
failing. It turned out to be because pulling the argocd image was
slow, timing out the test init. Now we ensure that the image is pulled
before performing `kubectl wait`, so that we only need to wait for the
pods to be ready.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>