chore: set defaults in testmain (#706)
* chore: set defaults in testmain

This keeps the default configuration in one place, and at the same time
allows for acceptance tests to not to have to use the make targets to
get the default values set (useful in e.g the IDE).

This also flips `USE_TESTCONTAINERS` to have a default of `true` (for now
we don't do that in CI). This is something which can be discussed to see
if that is desired by others.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>

* refactor: some minor tweaks

* add new line when printing
* inline `runTestSuite` logic

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>

---------

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>