fix: prevent deadlock
Fixes #710. While migrating from an older version using sdkv2 to
terraform-plugin-framework may need to recalculate the computed
attributes, it should definitely not deadlock in the meantime.

Added a test where we create a repository using `7.8.0`, and then
update it using the current version of this provider. Whereas before this
deadlocked, it now updates as intended, and once this initial migration
update has been done there is no state diff afterwards.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>