fix: set computed on namespace field
If `namespace` is not set the API will return `argocd` as the
namespace, since that's the default. To account for this we need to set
the namespace field as computed, which is also what the previous
implementation had.

We need to do something similar for `description`, where if we do not set
it in TF, we should not set a `description` value.

Fixes #783.

Also took the opportunity to fix some incorrect docs, which if needed can
be punted for another PR.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>