fix: add type to repository_credentials
Apart from that we add a bit of a hack since Argo CD doesn't return
type nor enable_oci when listing repository credentials. Pending a real
fix in Argo CD, in the meantime we will just trust that any repository
credential that has been created/updated with a type has been properly
saved, and we won't modify the state for those fields accordingly.

Fixes #791 (for now).

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>